refactor: Presenter가 View의 흐름을 결정하도록 변경
