+++ README.md
@@ -59,7 +59,7 @@
 - [x] MainActivity | commit 하는 부분 공통부분 메서드 분리
 - [x] MainActivity | Fragment 의 전역으로 미리 생성하도록 변경
 - [x] MainActivity | Fragment가 상태가 유지되도록 변경 (show, hide 사용)
-- [ ] HomeContract | navigate가 되는 책임 view에서만 담당하도록 변경
+- [x] HomeContract | navigate가 되는 책임 view에서만 담당하도록 변경
 - [ ] HomeFragment | 함수형 인터페이스로 변경
 - [ ] HomeFragment | 데이터 이동의 책임을 view -> presenter로 이동
 - [ ] HomeFragment | findViewById 삭제
