+++ app/src/main/java/woowacourse/movie/view/home/HomeContract.kt
@@ -1,6 +1,5 @@
 package woowacourse.movie.view.home
 
-import woowacourse.movie.domain.Showings
 import woowacourse.movie.view.home.movies.MovieUi
 
 interface HomeContract {
@@ -9,19 +8,6 @@ interface HomeContract {
     }
 
     interface View {
-        fun showMoviesScreen(
-            movieUis: List<MovieUi>,
-            navigate: (MovieUi) -> Unit,
-        )
-
-        fun navigateToReservation(
-            movieUi: MovieUi,
-            showings: Showings,
-        )
-
-        fun showTheaterSelectDialog(
-            movieUi: MovieUi,
-            navigate: (Showings) -> Unit,
-        )
+        fun showMoviesScreen(movieUis: List<MovieUi>)
     }
 }
