--- app/src/main/java/woowacourse/movie/view/home/movies/OnBottomSheetDialogEventListener.kt
@@ -1,7 +1,0 @@
-package woowacourse.movie.view.home.movies
-
-import woowacourse.movie.domain.Showings
-
-interface OnBottomSheetDialogEventListener {
-    fun onClick(showings: Showings)
-}
