+++ app/src/main/java/woowacourse/movie/view/home/theater/TheaterContract.kt
@@ -2,7 +2,7 @@ package woowacourse.movie.view.home.theater
 
 import woowacourse.movie.domain.Movie
 import woowacourse.movie.domain.Showings
-import woowacourse.movie.domain.Theater
+import woowacourse.movie.view.home.movies.MovieUi
 
 interface TheaterContract {
     interface Presenter {
@@ -12,15 +12,9 @@ interface TheaterContract {
     interface View {
         fun handleInvalidTicket()
 
-        fun showTheaterList(showings: List<Showings>)
-    }
-}
-
-class TheaterPresenter(
-    private val view: TheaterContract.View,
-) : TheaterContract.Presenter {
-    override fun fetchData(movie: Movie) {
-        val showings = movie.let { Theater.findTheatersShowingMovie(it.title) }
-        view.showTheaterList(showings)
+        fun showTheaterList(
+            showings: List<Showings>,
+            movieUi: MovieUi,
+        )
     }
 }
